requirements.txt
Django==4.0.3

djangorestframework==3.13.1

django_extensions==3.1.5
ipython
django-filter==22.1
